requirements.txt
asgiref==3.2.3
attrs==19.3.0
autobahn==19.11.2
Automat==0.8.0
cffi==1.13.2
constantly==15.1.0
cryptography==2.8
daphne==2.4.1
dj-database-url==0.5.0
Django==3.0.2
djangorestframework==3.11.0
gunicorn==20.0.4
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
Pillow==7.0.0
psycopg2-binary==2.8.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.19
PyHamcrest==1.9.0
pyOpenSSL==19.1.0
pytz==2019.3
service-identity==18.1.0
six==1.13.0
sqlparse==0.3.0
Twisted==19.10.0
txaio==18.8.1
whitenoise==5.0.1
zope.interface==4.7.1
